requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==7.0.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.7
cycler==0.12.1
distro==1.9.0
fonttools==4.54.1
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-genai==1.72.0
gTTS==2.5.4
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imbalanced-learn==0.14.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.9.2
narwhals==2.19.0
numpy==2.1.3
packaging==24.2
pandas==2.3.3
pillow==11.0.0
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pydeck==0.9.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.16.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==16.0
wheel==0.45.1
